Bump go to v1.23 (from v1.21) (#2132)
* Switch the go-setup actions to get the version from go.mod.

* Remove the go version steps from the sims workflow since the go setup action does that now.

* In Makefile: Add validate-go-version go the install target. Replace direct uses of the 'go' command with the $(GO) variable.

* Remove the explicit go cache steps from the sims workflow since go setup does that automatically now.

* Remove the explicit runsim install step since the make targets that need it are doing it on their own still anyway.

* Switch to go 1.23 in go.mod and update the Makefile to require go 1.23 (instead of 1.21).

* Update all the docker files to use golang:1.23-bullseye (from golang:1.21-bullseye).

* Remove internal/collections/maps.go since that functionality is in go by default now.

* Update the documentation to indicate go 1.23 is needed now.

* Add a changelog entry and create a summary.md that tells about the bump (since it's a pretty important change to note).

* Bump golangci-lint to v1.60.2.

* Fix all the new lint errors. Lots of integer casting problems were found.

* Fix TestIntBetween because I accidentally deleted a sort line when getting rid of our copy of the Keys function.

* In the heighliner job, try setting up go with the correct version. It's currently using v1.22 which causes make install to fail.

* Remove go setup from that job because it didn't help (kind of didn't expect it to anyway).

* In the heighliner job, try setting the BASE_VERSION env var to the needed golang version since the dockerfile might pick up on it.

* Remove the BASE_VERSION env var from the heighliner job since it's not the problem. Digging into their stuff, they don't yet have support for go v1.23, so we'll just have to wait.

* Include the golangci-lint bump in the changelog stuff.

+++ app/prefix.go
@@ -7,8 +7,8 @@ import (
 const (
 	AccountAddressPrefixMainNet = "pb"
 	AccountAddressPrefixTestNet = "tp"
-	CoinTypeMainNet             = 505
-	CoinTypeTestNet             = 1
+	CoinTypeMainNet             = uint32(505)
+	CoinTypeTestNet             = uint32(1)
 	Purpose                     = 44
 )
 
@@ -39,7 +39,7 @@ func SetConfig(testnet bool, seal bool) {
 	ConsNodePubKeyPrefix = AccountAddressPrefix + "valconspub"
 
 	config := sdk.GetConfig()
-	config.SetCoinType(uint32(CoinType))
+	config.SetCoinType(CoinType)
 	config.SetPurpose(Purpose)
 	config.SetBech32PrefixForAccount(AccountAddressPrefix, AccountPubKeyPrefix)
 	config.SetBech32PrefixForValidator(ValidatorAddressPrefix, ValidatorPubKeyPrefix)
